jamesdsp: Fix multi-arch issue with LOCAL_POST_INSTALL_CMD
LOCAL_POST_INSTALL_CMD runs once per arch.  On the first run, only
one of the built targets will be available and the second copy will
fail.  Depending on which order the targets are built, this may be
a spurious warning or a build failure.

Unfortunately, there does not seem to be a way to access the automatic
target specific make varibles (like $@) in LOCAL_POST_INSTALL_CMD
because kati escapes them all with backslashes.  So just put a test
for the target file's existence around the copy to fix the error.
This will result in one of the target files being copied twice, but
that is harmless.

Change-Id: I87d7a9665757c70da85970addbe63fab14713452